refactor form loading states
